Fix: (layouts/partials/footer.html) Alignment (#230)
* Fix: (layouts/partials/footer.html) Alignment

This removes the hard-coded padding-left property from these UL
elements.  It fixes the alignment in a narrow (e.g. smartphone)
display, and in a wide (e.g. desktop) display, the difference is
insignificant and no worse.

Fixes #192.  Thanks to Pamphile Roy (@tupui) for reporting.

* Fix: (assets/theme-css/styles.css) Margins for hero title, logo

Fixes #192.  Thanks to Pamphile Roy (@tupui) for reporting.

* Meta: (assets/theme-css/styles.css) Local variable

For indentation in Emacs's css-mode.  This way Emacs doesn't want to
change the indentation from the file's current 2-spaces setting.

* Change: (assets/theme-css/styles.css) "Hero" improvements

See <https://github.com/scientific-python/scientific-python-hugo-theme/pull/230>.

* Change: (assets/theme-css/styles.css) Logo on small screens

This enlarges the logo on screens above 400px wide, and on screens
that wide or narrower, the logo is displayed below the text.