Reduce flashing on page load
Currently, this is done by setting the background to the theme
preference (light/dark). This will stop working once we have a switch
to toggle between light and dark mode.

It was tempting to move the JS to executing earlier, but it needs to
be executed once the document elements are in place (for shortcuts
etc.).