Add: (render_shortcode_docs.py) Options processing
This allows each shortcode file to pass a dict of options to this
Python program.

Currently the only one supported is "render", which may be set to
False to disable rendering of the shortcode into the documentation.
This is useful when a shortcode will not render meaningfully into the
documentation file (e.g. the "toctree" shortcode only works when in an
appropriate file hierarchy).