Error for missing breadcrumbs (scientific-python/blog.scientific-python.org#158)
* layouts/partials/breadcrumbs.html: Log an error and fail the build
if a breadcrumb is missing (i.e. if a page used in a breadcrumb has a
missing "title" field in its front matter).

This only applies to content pages which are referred to in
breadcrumbs; other pages are still allowed to not have titles.

See <https://github.com/scientific-python/blog.scientific-python.org/issues/158>.