Enhancements for dropdowns: animations, open at initial state, and chevron indicators (#675)
This PR modifies the `dropdown` shortcode to add three new features, in
order to establish feature parity with the dropdowns from the [PyData
Sphinx
Theme](https://pydata-sphinx-theme.readthedocs.io/en/stable/user_guide/web-components.html#dropdowns)
and the
[sphinx-design](https://sphinx-design.readthedocs.io/en/pydata-theme/dropdowns.html)
projects. This was also discussed in
https://github.com/scientific-python/scientific-python-hugo-theme/issues/477#issuecomment-2037694735,
and should resolve that part of the conversation. Also, this PR is an
aftermath of #471.

In particular, the features added here are as follows:
- An "open" dropdown: the option to leave the dropdown open at initial
load, i.e., with the contents expanded instead of collapsed
- Chevron indicators: these are now displayed at the end of the
dropdown's title container when the dropdown has been expanded or is in
its collapsed state, and are animated, and there is also a choice
between
- Animations for the chevron indicators and for how the content is
displayed (fade-in, fade-in-slide-down) have been added, based on the
corresponding option for the directive

In addition to this, I've added several more examples to the shortcode's
docs: a success dropdown, a warning one, a danger one, a multi-line
example, one with code, and lastly, a nested dropdown.

I've checked for any accessibility issues/implications, and I think this
should be fine; I've closely followed the PST/SD implementation(s), and
the chevrons here work similarly.