Support named parameters for applicable shortcodes (#674)
This PR allows named parameters in the figure, button, and image
shortcodes and is an attempt to unify the base Hugo figure shortcode at
https://github.com/gohugoio/hugo/blob/a1cb15e1cf9b7606d36552dc31d2ade613318c55/tpl/tplimpl/embedded/templates/_shortcodes/figure.html
with the theme's shortcode (also for other applicable shortcodes).

The change here allows the shortcodes to be written in this manner: `{{<
figure src="image.jpg" alt="description" >}}`.

Based on https://gohugo.io/content-management/shortcodes/, the arguments
inside the opening tag are all using double-quoted strings, and I found
that using single-quoted strings can cause errors that are a little
difficult to debug based on Hugo's error messages, so I changed all
instances to use double-quoted strings where applicable to be on the
safer side.

We updated all the examples in the docs right above the shortcodes with
the updated syntax. Also, the link to the image of the cute puppies was
broken, so I added another one from Unsplash for our floofy friends :)

Additional context:

@goanpeca mentioned as a part of the work on
https://github.com/scientific-python-translations/ that this would be
nice to support, as Crowdin, as a part of the automations set up for the
projects' websites, only understands the named-parameters syntax for the
figure shortcode.

After the discussions (see below), we decided to expand the scope of
this PR to all shortcodes and to **not preserve backwards
compatibility** in favour of keeping just one syntax for the changed
shortcodes.

We have put together PRs that update to the new syntax for all projects that currently use the theme.

---------

Co-authored-by: Gonzalo Peña-Castellanos <goanpeca@gmail.com>
Co-authored-by: Stefan van der Walt <stefanv@berkeley.edu>